delete lsp and ipam with ip

+++ pkg/controller/pod.go
@@ -547,7 +547,7 @@ func (c *Controller) getPodKubeovnNets(pod *v1.Pod) ([]*kubeovnNet, error) {
 	return podNets, nil
 }
 
-func (c *Controller) changeVMSubnet(vmName, namespace, providerName, subnetName string, pod *v1.Pod) error {
+func (c *Controller) changeVMSubnet(vmName, namespace, providerName, subnetName string) error {
 	ipName := ovs.PodNameToPortName(vmName, namespace, providerName)
 	ipCr, err := c.config.KubeOvnClient.KubeovnV1().IPs().Get(context.Background(), ipName, metav1.GetOptions{})
 	if err != nil {
@@ -559,21 +559,12 @@ func (c *Controller) changeVMSubnet(vmName, namespace, providerName, subnetName 
 		// the returned pointer is not nil if the CR does not exist
 		ipCr = nil
 	}
-	if ipCr != nil {
-		if ipCr.Spec.Subnet != subnetName {
-			key := fmt.Sprintf("%s/%s", pod.Namespace, vmName)
-			if err := c.config.KubeOvnClient.KubeovnV1().IPs().Delete(context.Background(), ipName, metav1.DeleteOptions{}); err != nil {
-				if !k8serrors.IsNotFound(err) {
-					klog.Errorf("failed to delete ip %s, %v", ipName, err)
-					return err
-				}
-			}
-			klog.Infof("gc logical switch port %s", key)
-			if err := c.ovnLegacyClient.DeleteLogicalSwitchPort(key); err != nil {
-				klog.Errorf("failed to delete lsp %s, %v", key, err)
+	if ipCr != nil && ipCr.Spec.Subnet != subnetName {
+		if err := c.config.KubeOvnClient.KubeovnV1().IPs().Delete(context.Background(), ipName, metav1.DeleteOptions{}); err != nil {
+			if !k8serrors.IsNotFound(err) {
+				klog.Errorf("failed to delete ip %s, %v", ipName, err)
 				return err
 			}
-			c.ipam.ReleaseAddressByPod(key, subnetName)
 		}
 	}
 	return nil
@@ -637,7 +628,7 @@ func (c *Controller) handleAddPod(key string) error {
 		}
 		if isVmPod && c.config.EnableKeepVmIP {
 			pod.Annotations[fmt.Sprintf(util.VmTemplate, podNet.ProviderName)] = vmName
-			if err := c.changeVMSubnet(vmName, namespace, podNet.ProviderName, subnet.Name, pod); err != nil {
+			if err := c.changeVMSubnet(vmName, namespace, podNet.ProviderName, subnet.Name); err != nil {
 				klog.Errorf("change subnet of pod %s/%s to %s failed: %v", namespace, name, subnet.Name, err)
 				return err
 			}
